lpdiag: refactorization and improvements.
The file names changed in order to provide a clearer use. CL options changed also to improve usage. Info on actual working directory added, as well as check of readability of the required MPS file. The default run (i.e., without the CL params) uses one of small test MPS. Moreover, a number of improvements in comments, including removal of redundant comments.
